move useMathScope to redux area
