+++ client/src/features/sceneControls/mathItems/forms/Point/index.tsx
@@ -6,7 +6,8 @@ import {
 import React from "react";
 
 import FieldWidget, { useOnWidgetChange } from "../../FieldWidget";
-import { useMathErrors, useMathScope } from "../../mathScope";
+import { useMathScope } from "../../mathItemsSlice";
+import { useMathErrors } from "../../mathScope";
 import ItemTemplate from "../../templates/ItemTemplate";
 import { MathItemForm } from "../interfaces";
