+++ client/src/features/sceneControls/mathItems/mathItemsSlice/index.ts
@@ -3,6 +3,7 @@ import type { MathItemsState } from "./mathItems.slice";
 import mathItemsSlice, { actions, reducer } from "./mathItems.slice";
 
 export { select, actions, reducer };
+export * from "./hooks";
 
 export type { MathItemsState };
 export default mathItemsSlice;
